drf/requirements.txt
altgraph==0.17.2
aniso8601==9.0.1
asgiref==3.6.0
attrs==22.1.0
auto-py-to-exe==2.11.0
bottle==0.12.19
bottle-websocket==0.2.9
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
cryptography==37.0.4
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
dulwich==0.20.46
Eel==0.12.4
et-xmlfile==1.1.0
Faker==12.0.1
filelock==3.8.0
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==1.1.2
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
MarkupSafe==2.1.2
memory-profiler==0.60.0
mixer==7.2.2
more-itertools==8.14.0
msgpack==1.0.4
numpy==1.23.2
openpyxl==3.0.9
packaging==22.0
pefile==2021.9.3
pexpect==4.8.0
Pillow==9.0.0
pipenv==2022.12.19
pkginfo==1.9.2
platformdirs==2.5.2
poetry==1.3.1
poetry-core==1.4.0
poetry-plugin-export==1.2.0
promise==2.3
psutil==5.9.2
psycopg2==2.8.6
ptyprocess==0.7.0
pycparser==2.21
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.7
pywin32-ctypes==0.2.0
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.5
trove-classifiers==2022.12.22
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
